Added tests and moved code to tabular_dataset
