+++ google/cloud/aiplatform/datasets/tabular_dataset.py
@@ -15,15 +15,21 @@
 # limitations under the License.
 #
 
+import csv
 from typing import Optional, Sequence, Tuple, Union
 
 from google.auth import credentials as auth_credentials
+
+from google.cloud import bigquery
+from google.cloud import storage
 
 from google.cloud.aiplatform import datasets
 from google.cloud.aiplatform.datasets import _datasources
 from google.cloud.aiplatform import initializer
 from google.cloud.aiplatform import schema
 from google.cloud.aiplatform import utils
+
+from typing import List
 
 
 class TabularDataset(datasets._Dataset):
@@ -32,6 +38,140 @@ class TabularDataset(datasets._Dataset):
     _supported_metadata_schema_uris: Optional[Tuple[str]] = (
         schema.dataset.metadata.tabular,
     )
+
+    @property
+    def column_names(self) -> List[str]:
+        input_config = self._gca_resource.metadata.get("input_config")
+
+        if input_config is None:
+            raise RuntimeError("No input_config found for dataset")
+
+        gcs_source = input_config.get("gcs_source")
+        bq_source = input_config.get("bigquery_source")
+
+        if gcs_source:
+            gcs_source_uris = gcs_source.get("uri")
+
+            if gcs_source_uris and len(gcs_source_uris) > 0:
+                # Lexicographically sort the files
+                gcs_source_uris.sort()
+
+                # Get the first file in sorted list
+                return TabularDataset._retrieve_gcs_source_columns(
+                    self.project, gcs_source_uris[0]
+                )
+        elif bq_source:
+            bq_table_uri = bq_source.get("uri")
+            return TabularDataset._retrieve_bq_source_columns(bq_table_uri)
+
+        raise RuntimeError("No valid CSV or BigQuery datasource found.")
+
+    @classmethod
+    def _retrieve_gcs_source_columns(
+        cls, project_id: str, gcs_csv_file_path: str
+    ) -> List[str]:
+        """Retrieve the columns from a comma-delimited CSV file stored on Google Cloud Storage
+
+        Example Usage:
+
+            column_names = _retrieve_gcs_source_columns(
+                "gs://example-bucket/path/to/csv_file"
+            )
+
+            # column_names = ["column_1", "column_2"]
+
+        Args:
+            gcs_csv_file_path (str):
+                Required. A full path to a CSV files stored on Google Cloud Storage.
+                Must include "gs://" prefix.
+
+        Returns:
+            List[str]
+                A list of columns names in the CSV file.
+                
+        Raises:
+            RuntimeError: When the retrieved CSV file is invalid.
+        """
+
+        gcs_bucket, gcs_blob = utils.extract_bucket_and_prefix_from_gcs_path(
+            gcs_csv_file_path
+        )
+        client = storage.Client(project=project_id)
+        bucket = client.bucket(gcs_bucket)
+        blob = bucket.blob(gcs_blob)
+
+        # Incrementally download the CSV file until the header is retrieved
+        first_new_line_index = -1
+        start_index = 0
+        increment = 1000
+        line = ""
+
+        try:
+            while first_new_line_index == -1:
+                line += blob.download_as_bytes(
+                    start=start_index, end=start_index + increment
+                ).decode("utf-8")
+                first_new_line_index = line.find("\n")
+                start_index += increment
+
+            header_line = line[:first_new_line_index]
+
+            # Split to make it an iterable
+            header_line = header_line.split("\n")
+
+            csv_reader = csv.reader(header_line, delimiter=",")
+        except:
+            raise RuntimeError(
+                f"There was a problem extracting the headers from the CSV file at: { gcs_csv_file_path }"
+            )
+
+        return next(csv_reader)
+
+    @classmethod
+    def _retrieve_bq_source_columns(cls, bq_table_uri: str) -> List[str]:
+        """Retrieve the columns from a table on Google BigQuery
+
+        Example Usage:
+
+            column_names = _retrieve_bq_source_columns(
+                "bq://project_id.dataset.table"
+            )
+
+            # column_names = ["column_1", "column_2"]
+
+        Args:
+            bq_table_uri (str):
+                Required. A URI to a BigQuery table. 
+                Can include "bq://" prefix but not required.
+
+        Returns:
+            List[str]
+                A list of columns names in the BigQuery table.
+        """
+
+        # Remove bq:// prefix
+        prefix = "bq://"
+        if bq_table_uri.startswith(prefix):
+            bq_table_uri = bq_table_uri[len(prefix) :]
+
+        bq_source_components = bq_table_uri.split(".")
+        bq_project_id = bq_source_components[0]
+        bq_dataset_name = bq_source_components[1]
+        bq_table_name = bq_source_components[2]
+
+        query = f"""
+            SELECT column_name
+            FROM {bq_project_id}.{bq_dataset_name}.INFORMATION_SCHEMA.COLUMNS
+            WHERE table_name = '{bq_table_name}'
+        """
+
+        # Construct a BigQuery client object.
+        client = bigquery.Client(project=bq_project_id)
+        table = client.get_table(bq_table_uri)
+        table.schema()
+        query_job = client.query(query)  # Make an API request.
+
+        return [row.column_name for row in query_job]
 
     @classmethod
     def create(
