+++ google/cloud/aiplatform/training_jobs.py
@@ -15,7 +15,6 @@
 # limitations under the License.
 #
 
-import csv
 import datetime
 import functools
 import logging
@@ -29,7 +28,6 @@ from typing import Callable, Dict, List, Optional, NamedTuple, Sequence, Tuple, 
 
 import abc
 
-from google.cloud import bigquery
 from google.auth import credentials as auth_credentials
 from google.cloud.aiplatform import base
 from google.cloud.aiplatform import constants
@@ -2824,106 +2822,6 @@ class AutoMLTabularTrainingJob(_TrainingJob):
             sync=sync,
         )
 
-    # TODO: Add docstrings
-    @classmethod
-    def _retrieve_gcs_source_columns(
-        cls, project_id: str, gcs_csv_file_path: str
-    ) -> List[str]:
-        """Retrieve the columns from a comma-delimited CSV file stored on Google Cloud Storage
-
-        Example Usage:
-
-            column_names = _retrieve_gcs_source_columns(
-                "gs://example-bucket/path/to/csv_file"
-            )
-
-            # column_names = ["column_1", "column_2"]
-
-        Args:
-            gcs_csv_file_path (str):
-                Required. A full path to a CSV files stored on Google Cloud Storage.
-                Must include "gs://" prefix.
-
-        Returns:
-            List[str]
-                A list of columns names in the CSV file.
-                
-        Raises:
-            RuntimeError: When the retrieved CSV file is invalid.
-        """
-
-        gcs_bucket, gcs_blob = utils.extract_bucket_and_prefix_from_gcs_path(gcs_csv_file_path)
-        client = storage.Client(project=project_id)
-        bucket = client.bucket(gcs_bucket)
-        blob = bucket.blob(gcs_blob)
-
-        # Incrementally download the CSV file until the header is retrieved
-        first_new_line_index = -1
-        start_index = 0
-        increment = 1000
-        line = ""
-
-        try:
-            while first_new_line_index == -1:
-                line += blob.download_as_bytes(
-                    start=start_index, end=start_index + increment
-                ).decode("utf-8")
-                first_new_line_index = line.find("\n")
-                start_index += increment
-
-            header_line = line[:first_new_line_index]
-
-            # Split to make it an iterable
-            header_line = header_line.split("\n")
-
-            csv_reader = csv.reader(header_line, delimiter=",")
-        except:
-            raise RuntimeError(
-                "There was a problem extracting the headers from the CSV file."
-            )
-
-        return next(csv_reader)
-
-    @classmethod
-    def _retrieve_bq_source_columns(cls, bq_table_uri: str) -> List[str]:
-        """Retrieve the columns from a table on Google BigQuery
-
-        Example Usage:
-
-            column_names = _retrieve_bq_source_columns(
-                "bq://project_id.dataset.table"
-            )
-
-            # column_names = ["column_1", "column_2"]
-
-        Args:
-            bq_table_uri (str):
-                Required. A URI to a BigQuery table. 
-                Can include "bq://" prefix but not required.
-
-        Returns:
-            List[str]
-                A list of columns names in the BigQuery table.
-        """
-
-        bq_source_components = bq_table_uri.removeprefix("bq://").split(".")
-        bq_project_id = bq_source_components[0]
-        bq_dataset_name = bq_source_components[1]
-        bq_table_name = bq_source_components[2]
-
-        query = f"""
-            SELECT column_name
-            FROM {bq_project_id}.{bq_dataset_name}.INFORMATION_SCHEMA.COLUMNS
-            WHERE table_name = '{bq_table_name}'
-        """
-
-        # Construct a BigQuery client object.
-        client = bigquery.Client(project=bq_project_id)
-
-        query_job = client.query(query)  # Make an API request.
-
-        return [row.column_name for row in query_job]
-
     @base.optional_sync()
     def _run(
         self,
@@ -3020,31 +2918,9 @@ class AutoMLTabularTrainingJob(_TrainingJob):
         training_task_definition = schema.training_job.definition.automl_tabular
 
         if self._column_transformations is None:
-            # Retrieve all columns
-            gcs_source_files = dataset._gca_resource.metadata["inputConfig"][
-                "gcsSource"
-            ]["uri"]
-            bq_source = dataset._gca_resource.metadata["inputConfig"][
-                "bigquery_source"
-            ]["uri"]
-
-            column_transformations: List = []
-
-            if gcs_source_files and len(gcs_source_files) > 0:
-                # Lexicographically sort the files
-                gcs_source_files.sort()
-
-                # Get the first file in sorted list
-                column_names = AutoMLTabularTrainingJob._retrieve_gcs_source_columns(
-                    self.project, gcs_source_files[0]
-                )
-            elif bq_source:
-                column_names = AutoMLTabularTrainingJob._retrieve_bq_source_columns(
-                    bq_source
-                )
-
             column_transformations = [
-                {"AUTO": {"column_name": column_name}} for column_name in column_names
+                {"AUTO": {"column_name": column_name}}
+                for column_name in dataset.column_names
             ]
         else:
             column_transformations = self._column_transformations
